requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
requests==2.24.0
urllib3==1.25.10
python-telegram-bot==12.8
gTTS==2.1.1
gtts-token==1.1.3
pyttsx3==2.90
translator==0.0.9
langdetect==1.0.8
youtube-dl==2020.9.14
ffprobe==0.5
ffmpeg==1.4
googletrans==3.0.0
mysql-connector-python==8.0.21
